Generate CMOV instructions for a profitable subset of if-then blocks
In 541.leela benchmark, a performance improvement opportunity was identified by
transforming an if-then block into straight code using CMOV. This transformation
is profitable when the branch is not predicted very well during run-time.
In this change, a new pass called ForcedCMOVGeneration is added. This pass identifies
a candidate if-then block that has a single 'store' instruction in the 'then' block.
Once identified, a dummy variable is created on the stack and a 'select' instruction is used
to select between the actual address used in the 'then' block and the address of the dummy
variable on the stack. An example is shown here:
Before the pass:
  ....
  %cmp14 = icmp sgt i32 %5, %conv9
  br i1 %cmp14, label %if.then, label %if.end
if.then:
  store i32 %conv9, i32* %arrayidx.i35, align 4, !tbaa !2
  br label %if.end
if.end:
  .....
After the pass:
  ....
  %0 = alloca i32
  %cmp14 = icmp sgt i32 %5, %conv9
  %new_addr = select i1 %cmp14, i32* %arrayidx.i35, i32* %0
  store i32 %conv9, i32* %new_addr, align 4, !tbaa !2
  br label %if.end
if.end:
  .....
The Select instruction will be lowered to CMOV during ISelLowering stage.
This pass is a function pass that has been added to very end of LTO pass list.
This pass is triggered with -flto option during the linking stage.
Also, this pass is enabled by deafult without checking any target specific options.

Change-Id: I7996da84a3f89a778c6ca5ece1530b0042f57bc8
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/230269
Tested-by: Sudarsanam, Arvind <arvind.sudarsanam@intel.com>
Reviewed-by: Krishnaiyer, Rakesh <rakesh.krishnaiyer@intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>