[VPlan][2/4] Update VPlan CFGBuilder and CodeGen to use VPCallInstruction.
This patch updates VPlan's front-end and back-end components to
explicitly use VPCallInstruction specialization for Call instructions
in incoming IR. This allows us to replace any Call instruction specific
utilities or special handling performed in both CFG construction and
code-generator. This is the second patch in the series for the task
to centralize call vectorization decisions in VPlan.

Change-Id: I588cb2261b868f9b975490a6b61c83aa3bf6b075
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/228676
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Tested-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>