[SYCL] Temporary workaround in sycl program build for L0 plugin
Device libraries have been recently implemented in open-source SYCL
compiler. As a result the way how sycl program is built has changed.
piProgramCompile/piProgramLink calls are used to build the program. This
causes runtime error for L0 backend because these functions are not
implemented yet in L0 plugin. That is why temporary workaround was
implemented. It will be removed when piProgramCompile/piProgramLink will
be implemented in L0 plugin.

Change-Id: I23be3068d777912e85f5c6f2ce41e23360756940
Signed-off-by: Artur Gainullin <artur.gainullin@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/226549
Reviewed-by: Maslov, Sergey V <sergey.v.maslov@intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>