[VPlan][3/4] Introduce central analysis to determine call vectorization decisions.
This patch introduces a new VPlan-based analysis to compute
decisions on how VPCallInstructions should be handled in
outgoing vector code. These decisions are influenced by VF
and various external components like vector variants, target
library and call site/function attributes. The final set of
properties are explicitly recorded in the VPCallInstruction
which can be then used by other VPlan components down the
pipeline. This is the third patch in the series for the task
to centralize call vectorization decisions in VPlan.

Change-Id: Icff90a2e82710aaecefb5d7a2f2ee88582cef4b3
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/228677
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Tested-by: Karthik Senthil <karthik.senthil@intel.com>