[VPlan][LLVM CG] Serialize non-simple loads
The LIT test added in this patch used to crash because of discrepancy between DA
and CG decisions for a volatile load from uniform pointer. Previously, CG looked
into the uniform property of the pointer operand and decided to emit single load
only. However, DA marked the load as producing divergent value so later
getVectorValue on it resulted in crash.

Now we always produce vector value from the start (via serialization).

This change also does a similar change for atomic loads - they're never
vectorized now, only serialization is possible. However, for an atomic load from
a uniform address it's fine to keep a single load instead of VF
copies (loop-carried dependencies shouldn't exist and it's not a CG job to check
for/handle them).

Change-Id: Ibf802511467f541de0f71ece748dce34b07c1c8d
JIRA: CORC-6998
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/226829
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>