Ignore PHI only used in empty cleanup pads
This is an attempt at a better solution to the problem where SimplifyCFG
was introducing SSA violations while removing empty cleanup pads. My
previous solution avoided deleting any pads that contained lifetime
intrinsics, but it turns out that will be a very common occurance so I
think it's better to delete the empty pad and allow the lifetime end
intrinsic to be deleted.

I will also submit this patch to llorg, but I'm putting it in xmain
first because my previous change was worse than I expected it to be.

This addresses CMPLRLLVM-11251.

Change-Id: I7c901fc79da98fb78e58639ff9aa6ad3a6091d03
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/227038
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Lin Jr, Chang-Sun <chang-sun.lin.jr@intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>