[VPlan][4/4] Upgrade LLVM-IR vector codegen to use call vectorization decisions.
This patch simplifies VPlan's LLVM-IR vector codegen handling of
Call instructions by updating it to use CallVecProperties and
CallVecScenario which was appropriately determined during the call
vectorization decisions analysis. This change highlights the benefits
of centralizing call vectorization decisions in VPlan and how the
framework is leveraged in other VPlan components. This is the last
patch in the series for the task to centralize call vectorization
decisions in VPlan.

Change-Id: I5811c9b299fc45d6d5c2ce6c15c0d1add6676e91
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/228678
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Tested-by: Karthik Senthil <karthik.senthil@intel.com>