[SYCL] Fix the GPU driver version regexp
Linux:
19\.[4-9][3-9]\..*) -> 19\.(4[3-9]|[5-9]\d)\..*)
For example, this allows 19.50.*

Windows:
100\.(737[2-9]|73[7-9]\d|7[3-9]\d+|[7-9]\d+) -> \d+\.\d+\.100\.(737[2-9]|73[8-9]\d|7[4-9]\d+|[8-9]\d+)
This ensures that the first check for >100.7327 is not
nullified by the subsequent checks.

Finally, a pass-through of Windows through Linux checks was fixed by
explicitly breaking the regex via dot separators, exploting the fact
that this amounts to 3 blocks and 4 blocks for Linux and Windows
correpospondingly.

Change-Id: I1f3c4037cc11d1893bccd740157890ecb41460ce
Signed-off-by: Artem Gindinson <artem.gindinson@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/227772
Reviewed-by: Romanov, Vlad <vlad.romanov@intel.com>