[CMPLRLLVM-10533] Implement array use analysis for mcf.
This analysis keeps track of the range of the array that is being used after a
given array operation, expressed in terms of its operations. The goal of the
analysis for mcf is to prove that only the first 60 elements of the array will
be subsequently used, and so the sort only needs to sort up to that amount.

Change-Id: If83fef57e30a9f704f6016d51d59b4b256b73828
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/224734
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>