[CMPLRLLVM-11404][JumpThreading] Fix non-deterministic compilations
The reason for non-deterministic compiles was because data in
DenseMap<BasicBlock*, int> BlockThreadCount;
was carried through compilation of multiple functions.

For example, while threading FunctionA we created new basic block BB1.
Because threading across loop headers for newly created blocks is
disalowed we maintain a corresponding record in BlockThreadCount.
However, exactly in the same round of transformation this block was
deleted and memory for BB1 was freed. Later some transformation
(in my case it was LoopUnswitch) created new basic block BB2 in
FunctionBwhich (by accident) happened to have the same address as BB1 had.
Because the entry in BlockThreadCount that corresponds to BB1
was not removed, JumpThreading pass used it for processing BB2.

For baseline I observed 16 non-deterministic compilations
per 1000 iterations. After the change I see 0 mismatches.

Change-Id: Ibb2cad6831cc1edbd43b7dfd442c4116efbc69de
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/228128
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Lin Jr, Chang-Sun <chang-sun.lin.jr@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>