[VPlan][Predicator] Re-use predicate in case of full re-convergence
Copying the rest of the commit message from the actual comment in the sources...

Special case for predicate re-use in case of full re-convergence, i.e.
diamond pattern.

       \   /
        BB0 (D)
        / \
       /   \
      BB2  BB1
       \   /
        BB3  (Re-use BB0 predicate)

Note, that if BB3 would have another incoming edge we won't do the same
re-use and will fallback to the post-dominator frontier based algorithm.

This special case covers an important need of the region bypassing
algorithm that works post-linearization and needs to determine the region
boundaries.

Change-Id: I8a42db63da5bd490bf4533e9901126251e2e4703
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/229238
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>