[NFC] Update SubGroupBuiltinsVectorVariant.ll to match both before/after future change
We already have a TODO near the CHECK line, just update the check to allow an
"intermediate" flavour of the expected result. The future change for a "bad"
case would give the "spir_func" CallingConvention, but the "tail" property is
still future work. So, modify the CHECK to match all of the

  * "call" - current check, current behavior
  * "tail call spir_func" - current check, ideal
  * "call spir_func" - doesn't pass validation today, will be generated by a
    future patch.

Change-Id: Iebe89795227347dee13ddaa15741aad50dab47d1
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/229687
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: Mitropoulou, Konstantina <konstantina.mitropoulou@intel.com>
Reviewed-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>