Improve handling of ctors and dtors during target region filtering.
Special handling is needed since ctors and dtors have two routines
associated with them.  The offload entry contains the mangled name of
the "complete" ctor or dtor.  The target filtering checks the mangled
name in the offload entry and if it doesn't find an entry for that name
doesn't emit the function leaving the "base" ctor or dtor undefined.

This change will use the "complete" mangled name when looking for target
regions if none are found after using the "base" mangled name.

Change-Id: I3b2d2b6dcd882a3d580b55e8ef22457f5fd7a98a
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/230376
Reviewed-by: Yu, Jennifer <jennifer.yu@intel.com>
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>