[Driver][SYCL] Improve dependency file behaviors for -fintelfpga
This is a cherry-pick of https://github.com/intel/llvm/pull/1145

When compiling AOT for FPGA dependency files are generated that are used
by the aoc compilation.  Single step compilation is seamless as the dependency
file is generated then immediately used.  When compiling to object, keeping
track of the dependency file is not so intuitive.

To alleviate problems of not being able to find the dependency file, the
original dependency file is bundled up with the destination fat object and
when used is unbundled and passed to the aoc compilation.

NOTE:  A few small changes were needed for the Unbundler action to incorporate
changes that have only been done on intel/llvm that is dependent here.

Change-Id: I2d795b114c07adccc1ec87196b4fde6e3580ece1
Signed-off-by: Michael D Toguchi <michael.d.toguchi@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/234420
Reviewed-by: Gainullin, Artur <artur.gainullin@intel.com>
Tested-by: Gainullin, Artur <artur.gainullin@intel.com>