Propagate -cl-opt-disable and -g options to backend
Currenly for OpenCL kernels -g and -cl-opt-disable options are
propagated to backend compiler using opencl.compiler.options metadata.
For the scenario when SYCL compiler is used with CPU runtime this
solution doesn't work because SYCL kernel doesn't have this metadata.
Implemented additional propagation of these options to backend to
support the scenario when SYCL compiler is used.

Change-Id: Ib15928af57c6003e486a2476b6c9af683675dac0
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/230410
Reviewed-by: Gainullin, Artur <artur.gainullin@intel.com>
Tested-by: Gainullin, Artur <artur.gainullin@intel.com>