[DTrans] More preparation for opaque pointer handling
The DTrans tracking of types operated upon by certain call instructions
was keeping a list of PointerType objects that pointed to the actual
data type that was going to be operated upon by the call. The transform
passes would then get this list and invoke getPointerElementType() on it
to check the underlying type. Since getPointerElementType() is going away
with the change to opaque pointers, and the Transforms just need the element
type anyway, change the stored item to be the element type, and rename the
data structures and interfaces.

Change-Id: Id0edc64081421ba5d86d233a84bc5d286bf8dcda
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/230533
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>