[NFC][VPlan][1/2] Update vplan_hcfg_ssa_issues.ll to dump checks after merge loop exits transformation. Initially, the checks of this lit test were generated after the predicator. This lit test checks if we preserve SSA after loop massaging. After committing the following patch, this is not an issue any more:
commit a220530604d67f9b8cc08e8f6286ded78cb39617
[VPlan] Preserve SSA after merge loop exits transformation and while loop
canonicalization. This patch emits a phi node in NewLoopLatch for each value
that is defined in the loop body if this value is live-in in the loop header
or live out. If the basic block (that the value is defined) dominates all the
exiting blocks (e.g. LoopHeader), then we do not emit a phi node.

Change-Id: Ia036ff99d605722eefd9e29d46aa475c30ccb7fb
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/230173
Reviewed-by: Mitropoulou, Konstantina <konstantina.mitropoulou@intel.com>
Tested-by: Mitropoulou, Konstantina <konstantina.mitropoulou@intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>