[X86] Fix a dangling reference in debug printing code. Fix pass name in print messages.
The const MachineInstrBuilder &LEA was not extending the liftime
of the builder object since the last function called returns
a reference and not a temporary value.

This causes the LEA reference to dangle and then we try to access
the MachineInstr* by calling ->dump. Since the reference is
dangling, the MachineInstr* may be garbage or null leading to a
crash.

To fix this just save the MachineInstr * from the builder instead.

While there remove the #ifndef NDEBUG around the LEA variable and
cast to void instead.

Also fix the pass name in the print messages.

Change-Id: I0c36c4f711603014f8fb0aa1668c9ff65fcf4514
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/230602
Reviewed-by: Topper, Craig <craig.topper@intel.com>
Tested-by: Topper, Craig <craig.topper@intel.com>