[VPlan] Move DA calculation after simplifyPlainCFG()
Currently, DA is calculated just before simplifyPlainCFG(). simplifyPlainCFG()
consists of splitLoopsPreheader(), mergeLoopExits(),
singleExitWhileLoopCanonicalization() and splitLoopsExit().

MergeLoopExits() and singleExitWhileLoopCanonicalization() generate new phi
nodes. Currently, all these phi nodes are marked as divergent. However, this is
too conservative and might create performance regressions. In addition, the DA
should be updated because the shape of the loop changes and the instructions
might have different DA properties.

For example, in vplan_single_exit_while_loop.ll the exit condition of the inner
loop is marked as divergent but it is uniform. As a result, loop control flow
uniformity is triggered.

To avoid recalculating DA in these two transformations, this patch moves DA
calculation after simplifyPlainCFG().

After moving DA after simplifyPlainCFG(), the phi nodes that are generated by
merge loop exits and single exit while loop canonicalization were still marked
as divergent. This happens due to the fact that phi is erroneously marked as an
instruction with side effects. DA marks all the instructions with side effects
as divergent. This patch adds phi instruction to the group of the instructions
that do not have side effects. In this way, DA updates correctly the phi nodes
that are generated by merge loop exits transformation and single exits while
loop canonicalization. Unfortunately, I could not commit this change in a
separate patch because I could not replicate the problem that merge loop exits
transformation introduces. For the same reason, CMP, NOT, AllZeroCheck and
InductionInit.

Change-Id: I0c7ba03f7c4b8e0b543561eb329ae0d70985ee93
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/226951
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>