[CMPLRLLVM-11526][InstCombine] Ensure Descale preserves multi-user sums
Descale takes care to clone instructions before modifying them if they
have multiple users. When calling itself to descale a sum, the new
addition instruction must already exist so that multiple uses of the
addends are evident. Otherwise, Descale will modify what appears to be a
single use and incorrectly affect users of the original sum.

Creating/cloning the sum being descaled before calling Descale on its
addends is similar to what makeUniqueOp does. However, ::clone is not
quite suitable because of the case where the old sum is implemented as
an Or; in this case the new sum will be an Add.

Change-Id: I40ec831e875f94d071a3845a99f5d8de9f7e6b2c
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/230861
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>
Tested-by: Creech, Timothy M <timothy.m.creech@intel.com>