[VPlan][HIR] Do not use VPLoopEntity analysis results for reductions during CG.
VPLoopEntity analysis is done much earlier in VPlan pipeline and its
results are stale and can be incorrect when we reach codegen. This
patch modifies the codegen preparation pass to use def-use chains
in order to collect reduction related instructions, instead of
using LinkedVPValues from VPLoopEntity analysis. The def-use chain
is guaranteed to be up-to-date and hence the results are more
accurate before codegen.

Change-Id: I29f6621606c5fdafe5f84d12e040da2a6ef374e6
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/230910
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>