[X86] Move splat addends from the gather/scatter index operand to the base address.
This can avoid a vector add and a constant pool load. Or
an explicit broadcast in case of non-constant.

Also reverse the transform any time we encounter a constant
index addend that can't be moved to base. In that case pull
the constant from base into the index. This reduces code size
needed for the displacement since we needed the index add
anyway. Limit this to scale of 1 to avoid divisibility
and wrap issues.

Change-Id: I59fdaca60fe1973a8883483c4f838a7737275a98
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/233399
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Luo, Yuanke <yuanke.luo@intel.com>
Reviewed-by: Romanov, Vlad <vlad.romanov@intel.com>