CMPLRTOOLS-18828: Fix for support MSGL objects test cases
There is a chance that the rm instruction is hit before the Windows
releases the process. This issue could happen when check-lld or
check-all runs in parallel with many cores (32+). This fix moves
the "rm" instructions to the first thing to do when running the
test cases. The cleaning is used in case the tests need to be run
again (prevent incorrect libraries, etc.).

Change-Id: Ifa79a16e4ad264533e102930e44ec86e7b0b8306
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/231710
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>