CMPLRLLVM-11645: allow any local pointer values in OpenMP clauses.
This change-set relaxes restrictions on the values that can be
referenced inside OpenMP clauses. Global variable references
still must be represented as an optionally addrspacecast'ed
GlobaVariable's. Everything else is considered a reference
of a local variable.

We do our best to trace AllocaInst's from the clauses
to deduce the original item's type, but it seems to me
that it would be more robust if we had the type explicitly
represented in the clauses. VLA sizes may also be explicitly
referenced in the clauses, which would make our life much easier.

Change-Id: Ie36aee06614eb73df5e450dc70c61ff30071d194
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/231834
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>