CMPLRLLVM-11748 and CMPLRLLVM-11749: Fix for devirtualization after pulldown [XMAIN]
The community added a special check during LTO that checks how the user set the
the visibility. If the visibility flag is turned off the information for the
virtual calls in the module's global variables won't be updated. This disables
devirtualization. This changeset basically calls the function that updates the
global variables when we prove that whole program is safe.

Change-Id: I2534133aa1f3e8f23839515f62f3eda57e9db5d1
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/232251
Tested-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>