Don't limit ILPL parallelism when not using completion buffers
When not using completion buffers, (for example, when using
"-csa-ilpl-use-completion-buf=false",) ILPL parallelism is limited by
buffering on LICs. This is of course also not an unlimited resource, but
the boundaries are harder for the compiler to know. In this case we'll
leave it to the programmer to determine the limit.

Change-Id: I967f42a1bd8c757a64e9324aa32ad84b83b13f2d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/232472
Reviewed-by: Chen, Chih-Ping <chih-ping.chen@intel.com>
Tested-by: Creech, Timothy M <timothy.m.creech@intel.com>