[DTrans] Only capture offset value for DTrans element pointees when offset is not field boundary.
The change in https://git-amr-2.devtools.intel.com/gerrit/#/c/221449/ captured
byte offset values in the ElementPointee set to handle memintrinsics that
started within a structure padding location.

For cases where the offset was a field boundary the extra information was
redundant, and I found that it was incorrect in some cases. This information
was only used for dumps in these cases, and so is not necessary for these
cases.

This changes the tracking structure to use a union, so that it is known
whether the location is a field number or a byte offset, without relying
on magic numbers, so that the information in the dumps is correct.

Change-Id: I86320d566f4b71957fe06e4f7e2d43cc883da605
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/232846
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>