[FPGA] Runtime support for threads, mutexes and condition_variables
This is needed by one of our HLS tool features when they are packaged
in an opencl static library. It pushes everything that is problematic
in SPIR/OpenCL kernel IR into the runtime through opaque handles (just
as MSVC).
There is a thread,mutex and condition variable classes already in the
runtime, but is was never intended to be exposed inside kernel code.
This provides independent wrappers for IHC.

This is a fix for previous patch.

Change-Id: Ic4d86b6814ab4483df1e3613b12d887a1cc2c572
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/232767
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>
Tested-by: Dai, Chunyang <chunyang.dai@intel.com>
Reviewed-by: wuwei <wei.wu@intel.com>
Reviewed-by: Romanov, Vlad <vlad.romanov@intel.com>