[VPlan] Changes to support uniform control flow in HIR vector code generation
The changes here are an initial attempt to preserve uniform control flow
in HIR vector code generation. The changes handle uniform ifs by
emitting a label corresponding to the start of a VPBasicBlock and emitting
gotos at the end of a VPBasicBlock to appropriate successor block labels.
The gotos are placed in an if/else when the block has two successors.

Change-Id: I0d5ce757602fae85f1c9ee07d6b010d06bb3762b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/232975
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Tested-by: Guggilla, Satish <satish.guggilla@intel.com>