[OpenMP] Fix for front-end OpenMP entry block removal
This fixes the possible issue that alloc-hoisting can produce during
code extractor. The community assumes that the ArtificialEntry block
is empty, but we use that block to store the alloca. This changes
basically move those alloca instructions to the new entry block
that was created by the FE.

Related patches:
- https://reviews.llvm.org/D74372
- https://git-amr-2.devtools.intel.com/gerrit/#/c/127351/

Related issues:
- https://jira.devtools.intel.com/browse/CMPLRLLVM-387

Change-Id: I0a3b15bc589891c77a8fb80b06857026a7578842
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/234260
Tested-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Reviewed-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Reviewed-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>