Fix "target sections" construct on CSA so that kernel signature matches that on the host.
"Sections" is not lowered into a for-loop on CSA while it is on other
targets. Consequently, whereas on non-CSA targets the target call
has an argument for the loop upper bound (i.e. number of sections), such
argument is missing from the CSA version. This patch adds a dummy loop
upper bound for CSA as well so that the signatures match across all
platforms.

Change-Id: I0dcf4c42d05735e2540d9688cc34a1dc2f64b0aa
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/234465
Reviewed-by: Tian, Xinmin <xinmin.tian@intel.com>
Tested-by: Georgios Rokos <georgios.rokos@intel.com>
Reviewed-by: Dmitriev, Serguei N <serguei.n.dmitriev@intel.com>