[NFCI][VPlan] Do not use VPLoopEntity analysis results for live-out fixup during CG.
Currently live-outs from vector loop are fixed up by using VPLoopEntity
analysis data structures. This analysis is done much earlier in VPlan
pipeline and the results are stale and can be incorrect when we reach
codegen. This patch updates codegen to use explicit finalization
VPInstructions of the loop entities to fixup live-outs instead of using
an entity's LinkedVPValues list.

Change-Id: Iaf7b821b55745d5dacd5c4633590f9b46f224d62
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/234310
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Tested-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>