[SYCL][Driver] Improve the diagnostic for FPGA device link
This cherry-picks SYCL OS commit
https://github.com/intel/llvm/commit/6431be6780abc542a5a6732656a348b70a28756b
from SYCL OS PR
https://github.com/intel/llvm/pull/1077

FPGA compilations have a common use case of linking an object file against
an arhive that contains a pre-compiled device image. If the archive was
comprised without pulling in the object file in question, the presence
of any device kernel in the object file would signal incomplete device code
compilation/linkage.

Within the SYCL flow, a check for this case is performed by the
llvm-no-spir-kernel tool, so as to achieve a compile-time failure instead
of an obscure runtime failure due to a missing kernel.

This patch extends the Clang driver with the infrastructure to enhance
error messages upon receiving specific error codes from external tools.
The functionality is then used to enhance the diagnostic yielded by
llvm-no-spir-kernel in the aforementioned case.

Document the error-code handling infrastructure specifics

Change-Id: I3111d43f8e09b3108af42737ae412f6b5f3c4788
Signed-off-by: Artem Gindinson <artem.gindinson@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/234358
Reviewed-by: Romanov, Vlad <vlad.romanov@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Toguchi, Michael D <michael.d.toguchi@intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>