[SYCL] Update the kernel parameter rule to is-trivially-copy-constructible.
Since we really just want to be able to memcpy the type to the device,
'is-trivially-copyable' is not the correct trait. Since CWG1734, If we
want to support trivially copyable types, we would be required to create 1 of 4
different mechanisms for having a type on the device (depending on the
way the type is structured). Additionally, 2 of these ways require us to
ALSO have the type be default constructible.

This patch transitions to trivially-copy-constructible, so that we can
simply memcpy from the existing one into new memory.

Note: This is a cherry pick of intel/llvm PR 1144:
https://github.com/intel/llvm/pull/1144

Change-Id: I7f9c9a48b8ec165ac1ea88508ecbc7175eb34121
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/234397
Reviewed-by: Sergey Kanaev <sergey.kanaev@intel.com>
Tested-by: Keane, Erich <erich.keane@intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>