Fix the improperly remapped debug info
After the RemapInstruction, for some metadata representing debug
info, they are improperly remapped to a new created one, which
cause verification failure later. The cloned BasicBlocks reside in
the same function, so the source locations for the cloned
instructions and DbgVariableIntrinsic instruction's variables are
already correct, remap those to themselves to avoid verification
failure.

Change-Id: I8f3276e6e0d9e51726bee064452a56b23ae6fee5
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/234557
Reviewed-by: Wyma, Brock <brock.wyma@intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>
Tested-by: yifeihe <yifei1.he@intel.com>