[SYCL] Add a leaf limit to the execution graph (#1070)
This patch adds a leaf limit (per memory object) for the command
execution graph in order to avoid leaf bloat in applications that have
an overwhelming number of command groups that can be executed in
parallel. Limiting the number of leaves is necessary for reducing
performance overhead of regular cleanup of finished command nodes.

Whenever the limit is exceeded, the oldest leaf is added as a
dependency of the new one instead.

Signed-off-by: Sergey Semenov <sergey.semenov@intel.com>
(cherry picked from commit 7c293e24fe3a50cfcdddf4727800e7b9444f6a1a)

Change-Id: Ib2106cfb4d066eb954469ce4033a0337c013a6bd
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/234362
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Gainullin, Artur <artur.gainullin@intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>