[NFCI][VPlan] Force VPExternalUses to allow only PHINode in LLVM IR input
Before this patch the LCSSA form was already a requirement for which we
asserted in non-stress pipeline. This patch moves the restriction into
the pass manager by ensuring that LCSSA is scheduled before the VPlan even
for the stress-testing scenario.

It is an open question how to implement the same for the new pass manager.

Change-Id: I38ba9860260191c7176d772b3c93152bea3c3964
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/234942
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Tested-by: Elovikov, Andrei <andrei.elovikov@intel.com>