[VPlan][Predicator] Delay block splitting for not/and instructions creation
Predicate-related instructions (and/not/or) should be executed without the mask.
As such, in general, we need to have a separate block for them. Previously, the
block splitting to create such a new block was done at the time when such
instructions were created (after linearization during the predicates insertion).

The problem here is that we want to delay phis-to-blend processing until after
predicates are inserted and we don't want the phis to be modified prior to that
in order to carry original information in their incoming blocks. Unfortunately,
CFG modifications caused by the block splitting would update such phis too.

The solution to this problem is to make predicate-related instruction insertion
two-phase:
 - 1) Insert instructions *into* the blocks that are possibly predicated
 - do phis-to-blends processing
 - 2) Split the blocks from 1)

Another thing to note is that that delay has forced me to change the insertion
point of the "NOTs" generated in the predicator. Previously, such a NOT was
generated at the end of the basic block containing defining value (or preheader
for external definitions). Since now the same block is used to generate ANDs and
is splitted only later, I had to move the insertion point of the NOTs to right
after the def (and not the end of the block) to avoid putting the NOT into the
unpredicated block.

Change-Id: I5bb547ba5599de5b377510616904b7727c3c8f73
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/235636
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Tested-by: Elovikov, Andrei <andrei.elovikov@intel.com>