Fix the improperly remapped MD_loop metadata
For the loop latch BasicBlock inside UCF region, after the
CloneBasicBlock there is RemapInstruction function, inside which
the MD_loop metadata are improperly remapped to a new created one,
which cause verification failure later. The cloned BasicBlocks
reside in the same function, so the !llvm.loop metadata are
already correct, remap those to themselves to avoid verification
failure.
Change-Id: I02b807b0c8b30e5894c5b8f0698cbfd15a861bf8
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/236286
Reviewed-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>
Tested-by: yifeihe <yifei1.he@intel.com>
Reviewed-by: Mikhail Lychkov <mikhail.lychkov@intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>