[NFC][VPlan] Update tests to reduce diff in next patches
As part of that stop checking for some of the VPEntities internal dumps.
Checking for explicit VPInstructions is enough for the functional checks and we
want to limit the lifetime of the entities anyway.

Other changes are just moving the CHECK lines to right after the "define" to
mimic what the auto-update scripts do.

Change-Id: I9d4d7374c1bf0510879f618c71d9392c10a56714
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/236794
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>