[VPO][OpenMP]Drop clauses not supported by directive mapped from loop construct
The loop construct should be regarded as "parallel for" in FE. And FIRSTPRIVATE
and SHARED clauses are generated for loop construct by FE, but we have to drop
the clauses which are not supported by the mapped directive.
Here are two cases: if the loop construct is mapped to DIR_OMP_LOOP, the SHARED
clause is dropped. If it's mapped to DIR_OMP_SIMD, the FIRSTPRIVATE and SHARED
clauses are dropped.

JIRA: CMPLRLLVM-12256

Change-Id: Iba2a1a8202a21ccd2bf192fd0f31a6204ec00df5
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/236895
Reviewed-by: Zou, Feng <feng.zou@intel.com>
Tested-by: Zou, Feng <feng.zou@intel.com>