[OpenMP][Paropt] Change privatization/replacement logic for globals to only break expressions which reference those globals.
This is needed for an issue caused by replacement of a global with a
constexpr local, as previously, we were breaking and deleting
the constexpr local after the replacement of the original global.

However, this will also be a compile time (and potential performance)
improvement as we are not breaking all expressions, but only those
that need to be broken.

Change-Id: Ie5f5251fa2161a541f86e7493745ac7b89e7b340
JIRA: CMPLRLLVM-12250
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/237055
Reviewed-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Tested-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>