[SYCL] Remove hier_par test from clang tests
Test addresses regression in LowerWGScope pass. But:
  * Test doesn't check isolated LowerWGScope pass but depends on IR
  produced by FrontEnd.
  * LowerWGScope is llvm pass but located in the clang project. It
  should be moved to llvm project.

This issues are going to be addressed in https://github.com/intel/llvm.
LowerWGScope pass will be removed and test will be replaced by a lit
test in llvm project.

That is why delete hier_par test to avoid unstable behavior of this test
before pulldown.

Change-Id: Iade5ae8ef185cf8cf57408d97d93313f75861be2
Signed-off-by: Artur Gainullin <artur.gainullin@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/237087
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: Bader, Alexey <alexey.bader@intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>