Correctly handle AP-Int to bool cast kinds
CMPLRLLVM-18270 reported that a cast in a C language to a boolean type
(OpenCL is the easiest way to repro this) from an AP int caused an
assert.  The assert happened because the APInt code didn't properly
handle the cast and mistakenly made it an integral cast rather than
integer-to-boolean cast.  This patch corrects this.

Change-Id: Ib37add0fa38f64fa5910f137f8615b0f429fed3b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/237088
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Andrews, Elizabeth <elizabeth.andrews@intel.com>
Reviewed-by: Manna, Soumi <soumi.manna@intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>